more field property annotations
Mostly looking at fixing more implicit data race with more atomics. Some
of this (at the Julia level) is not strictly necessary since we already
intend to start using implicit release-consume ordering for references.
But since the user should not be changing these fields anyways, or risk
severe breakage, this should be fine to mark explicitly.